improve code style
