--- .gitignore
@@ -1,3 +1,0 @@
-/.idea/
-/components/bl-webcamPhoto-component/dist/
-
