+++ components/bl-webcamPhoto-component/src/buttons/done.js
