+++ components/bl-webcamPhoto-component/src/buttons/make-photo.js
