+++ components/bl-webcamPhoto-component/src/buttons/make-snapshot.js
