+++ components/bl-webcamPhoto-component/src/buttons/upload.js
