--- components/bl-webcamPhoto-component/src/utils/dataURLToBlob.js
@@ -1,5 +1,0 @@
-export async function dataURLToBlob(dataURL) {
-  const response = await fetch(dataURL)
-  
-  return response.blob()
-}
