+++ components/bl-webcamPhoto-component/src/utils/stop-user-media.js
@@ -1,0 +1,7 @@
+export const stopUserMedia = ref => {
+    if (ref.current.srcObject.stop) {
+        ref.current.srcObject.stop()
+    } else {
+        ref.current.srcObject.getTracks().forEach(track => track.stop())
+    }
+}
