+++ components/bl-webcamPhoto-component/src/utils/to-base64.js
@@ -1,0 +1,6 @@
+export const toBase64 = file => new Promise((resolve, reject) => {
+    const reader = new FileReader()
+    reader.readAsDataURL(file)
+    reader.onload = () => resolve(reader.result)
+    reader.onerror = error => reject(error)
+})
